Restructure rule code
  ref #65
